+++ changelog.md
@@ -7,6 +7,7 @@ case is still accepted), letters, and digits. **Single underscores** now separat
 * Breaking: The `EnvironmentVariableOverridePropertySource` has been removed. The standard `EnvironmentVariablesPropertySource` is now
 loaded by default, and takes precedence over other default sources just like the `EnvironmentVariableOverridePropertySource`
 did. To maintain similar behavior, configure it with a filtering `prefix`.
+* Add the ability to load a series of environment variables into arrays/lists via the `_n` syntax.
 
 ### 2.7.5